requirements.txt
Django==1.5.4
Pillow==2.5.1
django-classy-tags==0.5.1
django-sekizai==0.7
requests==2.3.0
simplejson==3.5.3
tweepy==2.3.0
wsgiref==0.1.2
